requirements.txt
beautifulsoup4
lucide-fasthtml
pillow
python-dotenv
python-fasthtml
requests
python-frontmatter
